+++ frontend/src/components/context-menu/account-settings-context-menu.tsx
@@ -1,6 +1,6 @@
-import { ContextMenu } from "./context-menu/context-menu";
-import { ContextMenuListItem } from "./context-menu/context-menu-list-item";
-import { ContextMenuSeparator } from "./context-menu/context-menu-separator";
+import { ContextMenu } from "./context-menu";
+import { ContextMenuListItem } from "./context-menu-list-item";
+import { ContextMenuSeparator } from "./context-menu-separator";
 import { useClickOutsideElement } from "#/hooks/useClickOutsideElement";
 
 interface AccountSettingsContextMenuProps {
